Migrating to micro builder construct
This is not tested yet, just the start. This still appears to keep working in the same way as before which is not too surprising. Eventually going to share this forward with the rest of the modules and then start work on simplifying the implementation along with the rest of the requirements